LayoutTests/svg/animations/animations-paused-when-removed-from-document-expected.txt

Tests that SVG animations are properly paused when removed from the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.areSVGAnimationsPaused is false
PASS svgElement.animationsPaused() is false
Removing SVG element from the document
PASS svgElement.animationsPaused() is true
Adding SVG element back into the document
PASS internals.areSVGAnimationsPaused is false
PASS svgElement.animationsPaused() is false
PASS successfullyParsed is true

TEST COMPLETE
